+++ supermercado.dal/src/supermercado/dal/dao/ProdutoDAO.java
@@ -7,7 +7,7 @@ import supermercado.dal.generics.JPAGenericDAO;
 import javax.enterprise.context.RequestScoped;
 
 @RequestScoped
-public class ProdutoDAO extends JPAGenericDAO<Produto, Short>
+public class ProdutoDAO extends JPAGenericDAO<Produto, Integer>
 implements IProdutoDAO{
 
 }
