+++ supermercado.dal/src/supermercado/dal/dao/interfaces/IProdutoDAO.java
@@ -3,6 +3,6 @@ package supermercado.dal.dao.interfaces;
 import supermercado.dal.entidade.*;
 import supermercado.dal.generics.*;
 
-public interface IProdutoDAO extends IGenericDAO<Produto, Short>{
+public interface IProdutoDAO extends IGenericDAO<Produto, Integer>{
 
 }
